@@ -1,10 +1,19 @@
 module github.com/chill/plaidqif
 
-go 1.16
+go 1.19
+
+require (
+	github.com/plaid/plaid-go v1.10.0
+	gopkg.in/alecthomas/kingpin.v2 v2.2.6
+)
 
 require (
 	github.com/alecthomas/template v0.0.0-20190718012654-fb15b899a751 // indirect
-	github.com/alecthomas/units v0.0.0-20210208195552-ff826a37aa15 // indirect
-	github.com/plaid/plaid-go v0.0.0-20210108171133-f3ad50a114c7
-	gopkg.in/alecthomas/kingpin.v2 v2.2.6
+	github.com/alecthomas/units v0.0.0-20211218093645-b94a6e3cc137 // indirect
+	github.com/golang/protobuf v1.5.2 // indirect
+	github.com/stretchr/testify v1.8.0 // indirect
+	golang.org/x/net v0.2.0 // indirect
+	golang.org/x/oauth2 v0.2.0 // indirect
+	google.golang.org/appengine v1.6.7 // indirect
+	google.golang.org/protobuf v1.28.1 // indirect
 )
